backend/requirements.txt
Flask==2.3.2
Flask-Cors==3.0.10
Flask-PyMongo==2.3.0
python-dotenv==1.0.0
Werkzeug==2.3.6
bcrypt==4.0.1
PyJWT==2.8.0
Pillow==9.5.0
openai==1.14.3
tensorflow==2.17.1

